backend/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
branca==0.7.1
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtradeapicall==1.0.20
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dateparser==1.2.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
folium==0.15.1
fonttools==4.48.1
greenlet==3.0.3
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwt==1.3.1
kaleido==0.2.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mistune==3.0.2
mutagen==1.46.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.17.1
scapy==2.5.0
seaborn==0.13.2
shelved-cache==0.3.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tabulate==0.8.10
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
virtualenv==20.25.0
wbdata==1.0.0
wbgapi==1.0.12
Werkzeug==3.0.1
WTForms==3.1.2
xyzservices==2023.10.1
